pyproject.toml
[project]
name = "mlops-with-databricks"
version = "0.0.1"
description = "MLOps with Databricks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "lightgbm>=4.5.0, <5",
    "scikit-learn>=1.5.1, <2",
    "cloudpickle>=3.0.0, <4",
    "mlflow>=2.16.0, <3",
    "numpy>=1.26.4, <2",
    "pandas>=2.2.2, <3",
    "pyarrow>=15.0.2, <16",
    "cffi>=1.17.1, <2",
    "scipy>=1.14.1, <2",
    "matplotlib>=3.9.2, <4",
    "databricks-feature-engineering>=0.6, <1",
    "pre-commit>=4.0.1",
    "loguru>=0.7.2",
    "build>=1.2.2.post1",
]

[project.optional-dependencies]
dev = [
    "databricks-connect>=15.4.1, <16",
    "databricks-sdk>=0.32.0, <0.33",
    "ipykernel>=6.29.5, <7",
    "pip>=24.2"
]

[tool.ruff]
line-length = 120
select = [
    "F",  # pyflakes rules
    "E",  # pycodestyle error rules
    "W",  # pycodestyle warning rules
    "B",  # flake8-bugbear rules
    "I",  # isort rules
]
ignore = [
    "E501",  # line-too-long
]

[tool.ruff.format]
indent-style = "space"

[project.scripts]
hello = "mlops_with_databricks:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
